image

+++ colour_helper.py
@@ -57,6 +57,6 @@ print(is_light(white_pixel))      # True
 
 def pixel_to_grayscale(pixel: tuple) -> tuple:
     """Returns a grayscale version of the given pixel"""
-    gray = pixel[0] * 0.3 + pixel[1] * 0.59 + pixel[2] * 0.11
+    gray = int(pixel[0] * 0.3 + pixel[1] * 0.59 + pixel[2] * 0.11)
 
     return (gray, gray, gray)
